# docs/openapi.json
{
  "components": {
    "schemas": {
      "BatchReq": {
        "properties": {
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "BatchReq",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "InNetwork API",
    "version": "3.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/geocode": {
      "get": {
        "description": "Geocode a free-text `q`, or a structured {postalcode, city, state}. The\nfrontend sends the structured form for the map center, so accept both rather\nthan 422-ing on a missing `q`.",
        "operationId": "api_geocode_api_geocode_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "postalcode",
            "required": false,
            "schema": {
              "default": "",
              "title": "Postalcode",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "city",
            "required": false,
            "schema": {
              "default": "",
              "title": "City",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "default": "",
              "title": "State",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Api Geocode Api Geocode Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Geocode"
      }
    },
    "/api/geocode/batch": {
      "post": {
        "operationId": "api_geocode_batch_api_geocode_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Api Geocode Batch Api Geocode Batch Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Geocode Batch"
      }
    },
...
    "/api/insurance/{npi}": {
      "get": {
        "description": "Coverage for one NPI. Estimated-tier plans need a `state` to resolve;\nwithout it only verified answers appear.",
        "operationId": "insurance_for_api_insurance__npi__get",
        "parameters": [
          {
            "in": "path",
            "name": "npi",
            "required": true,
            "schema": {
              "title": "Npi",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "default": "",
              "title": "State",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Insurance For Api Insurance  Npi  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Insurance For"
      }
    },
    "/api/npi": {
      "get": {
        "operationId": "api_npi_api_npi_get",
        "parameters": [
          {
            "in": "query",
            "name": "zip",
            "required": false,
            "schema": {
              "default": "",
              "title": "Zip",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "city",
            "required": false,
            "schema": {
              "default": "",
              "title": "City",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "default": "",
              "title": "State",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "npi",
            "required": false,
            "schema": {
              "default": "",
              "title": "Npi",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "name",
            "required": false,
            "schema": {
              "default": "",
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "taxonomy",
            "required": false,
            "schema": {
              "default": "",
              "title": "Taxonomy",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "default": "",
              "title": "Type",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Api Npi Api Npi Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Npi"
      }
    },
    "/api/providers/search": {
      "get": {
        "description": "One call: query NPPES, attach insurance flags (verified + estimated), then \u2014\nfor a radius search \u2014 geocode the candidate pool, keep only providers truly\nwithin `radius` miles of the ZIP center, sort by distance, and return ranked\nresults with coordinates already attached. The backend is authoritative for the\nradius boundary; the frontend plots the returned pins without re-geocoding.",
        "operationId": "providers_search_api_providers_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "zip",
            "required": false,
            "schema": {
              "default": "",
              "title": "Zip",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "city",
            "required": false,
            "schema": {
              "default": "",
              "title": "City",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "default": "",
              "title": "State",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "npi",
            "required": false,
            "schema": {
              "default": "",
              "title": "Npi",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "name",
            "required": false,
            "schema": {
              "default": "",
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "taxonomy",
            "required": false,
            "schema": {
              "default": "",
              "title": "Taxonomy",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "default": "",
              "title": "Type",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Miles from the ZIP center to keep; 0 disables distance filtering",
            "in": "query",
            "name": "radius",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Miles from the ZIP center to keep; 0 disables distance filtering",
              "title": "Radius",
              "type": "integer"
            }
          },
          {
            "description": "Comma-separated plan ids the provider must accept, e.g. 'medicare,aetna'",
            "in": "query",
            "name": "accepts",
            "required": false,
            "schema": {
              "default": "",
              "description": "Comma-separated plan ids the provider must accept, e.g. 'medicare,aetna'",
              "title": "Accepts",
              "type": "string"
            }
          },
          {
            "description": "'verified' = require confirmed acceptance; 'any' = also allow estimated matches",
            "in": "query",
            "name": "accepts_mode",
            "required": false,
            "schema": {
              "default": "verified",
              "description": "'verified' = require confirmed acceptance; 'any' = also allow estimated matches",
              "title": "Accepts Mode",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "geocode",
            "required": false,
            "schema": {
              "default": true,
              "title": "Geocode",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Providers Search Api Providers Search Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Providers Search"
      }
    },
    "/api/reverse": {
      "get": {
        "operationId": "api_reverse_api_reverse_get",
        "parameters": [
          {
            "in": "query",
            "name": "lat",
            "required": true,
            "schema": {
              "title": "Lat",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "lon",
            "required": true,
            "schema": {
              "title": "Lon",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Api Reverse Api Reverse Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Reverse"
      }
    },
...
  }
}